refactor: restructure css files and componenets
