+++ src/app/_components/favoriteIcon/FavoriteIcon.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import { Star } from 'lucide-react'
+import './favoriteIcon.scss'
+
+export default function FavoriteIcon({ isFavorite }: { isFavorite: boolean }) {
+  return <Star className="favorite-icon" data-active={isFavorite} />
+}
