+++ src/app/_components/posters/posters.scss
@@ -4,7 +4,7 @@
   display: flex;
 }
 
-.posters-column {
+.posters__column {
   padding: calc(var(--unit-3) / 2);
   width: 20%;
   gap: var(--unit-3);
@@ -18,6 +18,7 @@
   position: relative;
   background-color: var(--black);
   overflow: hidden;
+  transition: transform 0.1s ease, filter 4s ease;
 
   &::after {
     opacity: 0;
@@ -35,12 +36,14 @@
   &:hover {
     background-color: var(--black);
     z-index: 1;
+    transform: scale(1.04);
+    filter: hue-rotate(360deg);
 
     &::after {
       opacity: 1;
     }
 
-    .poster__icon {
+    .favorite-icon {
       color: var(--pink);
     }
   }
@@ -52,19 +55,4 @@
 
 .poster__img {
   width: 100%;
-}
-
-.poster__icon {
-  color: transparent;
-  z-index: 1;
-  position: absolute;
-  bottom: var(--unit-4);
-  right: var(--unit-4);
-  transition: color 0.2s ease-in, transform 0.2s ease-in;
-  transform: scale(0.8);
-
-  &--active {
-    fill: var(--pink);
-    transform: scale(1.1);
-  }
 }
