+++ src/app/_styles/globals.css
@@ -1,37 +1,11 @@
 @import 'tailwindcss';
 @import './variables.css';
 
-body {
-  background: var(--red);
-  font-family: var(--font-lacquer);
-  padding: 0 240px;
-}
-
-.container {
-  margin-top: var(--unit-lg);
-  max-width: 100rem;
-}
-
-.page-frame {
-  padding: 10px;
-}
-
-.outer {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  background: var(--red-dark);
-  padding: calc(var(--unit-lg) - var(--unit-1));
-  padding-top: 0;
-  min-height: 100vh;
-}
-
-.portrait-image {
-  color: transparent;
-  position: fixed;
-  bottom: 0;
-  right: -100px;
-  z-index: 2;
-  pointer-events: none;
-  max-width: 40%;
-}
+/* remove tailwind */
+/* update classnames to be kebab case */
+/* update folder structure */
+/* add metrics */
+/* look into performance */
+/* look into PPR */
+/* ask ai if any obvious fixes */
+/* generate readme with diagram */
